authentication-service/requirements.txt
anyio==4.6.2.post1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
autopage==0.5.2
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cliff==4.7.0
cmd2==2.5.0
colorama==0.4.6
cryptography==43.0.3
debtcollector==3.0.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
graphql-core==3.2.5
h11==0.14.0
idna==3.10
ipython==8.12.3
iso8601==2.1.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jwt==1.3.1
keystoneauth1==5.8.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
msgpack==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
netaddr==1.3.0
netifaces==0.11.0
os-service-types==1.7.0
oslo.i18n==6.4.0
oslo.serialization==5.5.0
oslo.utils==7.3.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pbr==6.1.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.6
prettytable==3.11.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyotp==2.9.0
pyparsing==3.2.0
pyperclip==1.9.0
pyreadline3==3.5.4
python-barbicanclient==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==8.0
redis==5.1.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.0
stevedore==5.3.0
strawberry-graphql==0.246.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
yarg==0.1.9
